refactor(tools): share formatter feedback helpers
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
