+++ src/voidcode/tools/multi_edit.py
@@ -6,13 +6,16 @@ from typing import ClassVar
 from pydantic import ValidationError
 
 from ..hook.config import RuntimeHooksConfig
-from ._formatter import FormatterExecutionResult, FormatterExecutor
+from ._formatter import (
+    FormatterExecutionResult,
+    FormatterExecutor,
+    formatter_diagnostics,
+    formatter_payload,
+)
 from ._pydantic_args import MultiEditArgs
 from .contracts import ToolCall, ToolDefinition, ToolResult
 from .edit import (
     EditTool,
-    formatter_diagnostics,
-    formatter_payload,
     read_utf8_text,
     summarize_diff,
 )
